acpi: build QEMU table for PPI virtual memory device
To avoid having to hard code the base address of the PPI virtual memory
device we introduce a QEMU ACPI table that holds the base address, if a
TPM 1.2 or 2 is used. This table gives us flexibility to move the base
address later on. We also write the size of the device's address space
into the table.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>