tpm_tis_common: Remove check for active locality on INT_ENABLE & INT_STATUS
The TPM TIS I2C spec describes in the table in section "Interface Locality
Usage per Register" that the TPM_INT_ENABLE and TPM_INT_STATUS registers
must be writable for any locality even if the locality is not the active
locality. Therefore, remove the checks whether the writing locality is the
active locality for these registers.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>