tpm: Use 16kb size for TPM CRB romd device read-only memory area
The romd device's read-only memory area must be at least the size of a
page of the host for it to be mapped into memory. Page sizes on aarch64
can be 4kb, 16kb, or 64kb. If a smaller size is used then crashes will
occur (memory was not mapped) or the VM will not work properly after
migration to the destination host. Therefore, choose 16kb as the size
of the memory area to map, even though only 4kb (0x1000) are needed per
TPM_CRB_ADDR_SIZE.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>